Fix git import for packages
This fixes importing for packages coming from git.  It adds support
for importing over ssh and gets rid of dependencies on sed ( because
OSX's sed is trash ) and awk.

The only thing not supported is packages that want to use the
Git daemon protocol, but it's recommended that you wrap that in
ssh/https anyways, due to it's concerning lack of auth.